+++ include/sdf/Error.hh
@@ -88,6 +88,9 @@ namespace sdf
     /// \brief A filesystem directory does not exist.
     DIRECTORY_NONEXISTANT,
 
+    /// \brief A model with an invalid canonical link.
+    MODEL_CANONICAL_LINK_INVALID,
+
     /// \brief A model without a link.
     MODEL_WITHOUT_LINK,
 
@@ -96,6 +99,25 @@ namespace sdf
 
     /// \brief A link has invalid inertia.
     LINK_INERTIA_INVALID,
+
+    /// \brief A joint has an invalid child link.
+    JOINT_CHILD_LINK_INVALID,
+
+    /// \brief A joint has an invalid parent link.
+    JOINT_PARENT_LINK_INVALID,
+
+    /// \brief A joint has the same link specified as parent and child.
+    JOINT_PARENT_SAME_AS_CHILD,
+
+    /// \brief The frame attached-to value does not match the name of an
+    /// existing frame in the current scope.
+    FRAME_ATTACHED_TO_INVALID,
+
+    /// \brief The frame attached-to graph contains a cycle.
+    FRAME_ATTACHED_TO_CYCLE,
+
+    /// \brief The frame attached-to graph has an internal error.
+    FRAME_ATTACHED_TO_GRAPH_ERROR,
 
     /// \brief Indicates that reading an SDF string failed.
     STRING_READ,